project_expense_tracking_system/requirements.txt
streamlit==1.45.1
pandas==2.2.3
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
mysql-connector-python==9.3.0
requests==2.32.3
pytest==8.3.5